[FEAT]#82 DELETE pages/qa 및 _prod/archives 생성
