--- src/pages/QA/index.stories.tsx
@@ -1,22 +1,0 @@
-import { ComponentMeta, ComponentStory } from '@storybook/react';
-import QA from '.';
-
-// More on default export: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
-export default {
-  title: 'PAGE/QA',
-  component: QA,
-
-  // More on argTypes: https://storybook.js.org/docs/react/api/argtypes
-  argTypes: {},
-} as ComponentMeta<typeof QA>;
-
-// More on component templates: https://storybook.js.org/docs/react/writing-stories/introduction#using-args
-const Template: ComponentStory<typeof QA> = (args) => <QA {...args} />;
-
-export const Primary = Template.bind({});
-// More on args: https://storybook.js.org/docs/react/writing-stories/args
-Primary.args = {};
-
-Primary.parameters = {
-  layout: 'fullscreen',
-};
