+++ src/pages/_prod/archives/[archive_id].tsx
@@ -1,0 +1,30 @@
+import { GetServerSideProps } from 'next';
+import { useRouter } from 'next/router';
+import { useState } from 'react';
+import { ArchiveTemplate } from '../../../components/prototypes/ArchiveTemplate';
+import { useProjectList } from '../../../hooks/queries/productQuery';
+import { ModeType } from '../../../types/common/propsTypes';
+
+const ArchivePage = () => {
+  const router = useRouter();
+  const [mode] = useState<ModeType>('dark');
+  const { archive_id, page } = router.query;
+  const { productList, totalCount } = useProjectList({
+    archive: archive_id,
+    page,
+  });
+
+  // if (isProjectListLoading) return <LayOut mode={mode} />;
+  return <ArchiveTemplate mode={mode} productList={productList} totalCount={totalCount}></ArchiveTemplate>;
+};
+export default ArchivePage;
+
+export const getServerSideProps: GetServerSideProps = async (context) => {
+  const { query } = context;
+  const { archive_id } = query;
+  return {
+    props: {
+      archive_id,
+    },
+  };
+};
